% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cori_network_communities.R
\docType{data}
\name{cori_network_communities}
\alias{cori_network_communities}
\title{Cori Network Communities}
\source{
https://github.com/ruralinnovation/map48-project
}
\value{
\item{cori_network_communities}{a tibble}
}
\description{
Sample cori network community data with 4 columns - rin_community, state_abbr, county, geoid_co.
}
\examples{
data(cori_network_communities)
str(cori_network_communities)
}
\keyword{cori_network_communities}
