% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{cori_primary}
\alias{cori_primary}
\alias{cori_orange}
\alias{cori_blue}
\alias{cori_purple}
\title{CORI palettes}
\format{
An object of class \code{character} of length 7.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.
}
\source{
CORI Data Visualization Style Guide
}
\usage{
cori_primary

cori_orange

cori_blue

cori_purple
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the CORI Data Visualization Style Guide.
}
\concept{cori palettes}
\keyword{datasets}
