% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_threshold_annotate.R
\name{geom_threshold_annotate}
\alias{geom_threshold_annotate}
\title{geom_threshold_annotate}
\usage{
geom_threshold_annotate(intercept, axis = "x", label = "", shift = 0)
}
\arguments{
\item{intercept}{x or y intercept (indicated by axis) through which the vertical or horizontal threshold is made.}

\item{axis}{Axis along which a threshold is to be added. Default is "x" indicating a vertical threshold}

\item{label}{Label to add to the graph. Defaults to an empty string}

\item{shift}{Variable to shift the label by "shift" units above thresholding line along "y", or by "shift" units to the right of thresholding line along "x"}
}
\description{
This is a utility function that allows the user to add a line threshold with text annotation using an intercept parameter.
The shift parameter allows the user to relatively adjust the text with respect to the defined thresholding.
}
