% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_sample_data.R
\docType{data}
\name{population_sample_data}
\alias{population_sample_data}
\title{Population Sample Data}
\source{
https://github.com/ruralinnovation/map48-project
}
\value{
\item{population_sample_data}{a tibble}
}
\description{
Sample population data with 54 columns.
}
\examples{
data(population_sample_data)
str(population_sample_data)
}
\keyword{population_sample_data}
