% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_work_by_ed_level.R
\docType{data}
\name{remote_work_by_ed_level}
\alias{remote_work_by_ed_level}
\title{Remote work by Education Level}
\source{
https://github.com/ruralinnovation/map48-project
}
\value{
\item{remote_work_by_ed_level}{a tibble}
}
\description{
Table with long form version of the time_series_sample_data table. There are
3 columns - date, education_level and pct_working_remotely
}
\examples{
data(remote_work_by_ed_level)
str(remote_work_by_ed_level)
}
\keyword{remote_work_by_ed_level}
