% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rural_urban_2021.R
\docType{data}
\name{rural_urban_2021}
\alias{rural_urban_2021}
\title{Time Series Sample Data}
\source{
https://github.com/ruralinnovation/map48-project
}
\value{
\item{rural_urban_2021}{a tibble}
}
\description{
Sample time series data with a date column and 4 other columns indicating
percentage of remotely working individuals across 4 categories
}
\examples{
data(rural_urban_2021)
str(rural_urban_2021)

}
\keyword{rural_urban_2021}
