% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_cori.R
\name{scale_color_cori}
\alias{scale_color_cori}
\title{Cori color scales}
\usage{
scale_color_cori(color = cori_primary)
}
\arguments{
\item{color}{Pass a cori color palette. Deafults to \code{cori_primary} colors}
}
\description{
This is a description
}
