% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_cori_geom_defaults.R
\name{set_cori_geom_defaults}
\alias{set_cori_geom_defaults}
\title{The CORI \link{ggplot2} theme}
\usage{
set_cori_geom_defaults(base_family = "TT Hoves")
}
\arguments{
\item{base_family}{The base font family for the theme.}
}
\description{
\code{set_cori_geom_defaults} provides a \link{ggplot2} theme formatted according to the
CORI style guide, with sensible defaults.
}
