% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cori.R
\name{theme_cori}
\alias{theme_cori}
\title{A \link{ggplot2} theme formatted in the CORI style}
\usage{
theme_cori(base_family = "TT Hoves", base_size = 10.5)
}
\arguments{
\item{base_family, base_size}{base font family and size}
}
\description{
\code{theme_cori} provides a \link{ggplot2} theme formatted according to the
CORI style guide for web, with sensible defaults.
}
