% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Thresholding}
\usage{
threshold(intercept, axis = "x", label = "", shift = 0)
}
\arguments{
\item{intercept}{x or y intercept (indicated by axis) through which the vertical or horizontal threshold is made.}

\item{axis}{Axis along which a threshold is to be added. Default is "x" indicating a vertical threshold}

\item{label}{Label to add to the graph. Defaults to an empty string}

\item{shift=0}{Pass a cori color palette. Deafults to \code{cori_primary} colors}
}
\description{
This is a description
}
