% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_sample_data.R
\docType{data}
\name{time_series_sample_data}
\alias{time_series_sample_data}
\title{Time Series Sample Data}
\source{
https://github.com/ruralinnovation/map48-project
}
\value{
\item{time_series_sample_data}{a tibble}
}
\description{
Sample time series data with a date column and 4 other columns indicating
percentage of remotely working individuals across 4 categories
}
\examples{
data(time_series_sample_data)
str(time_series_sample_data)

}
\keyword{time_series_sample_data}
