% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cori.R
\docType{data}
\name{cori_font}
\alias{cori_font}
\title{TT Hoves font name R variable aliases}
\format{
length 1 character vector
}
\usage{
cori_font
}
\description{
\code{cori_font} == "\verb{TT Hoves}"
}
\keyword{datasets}
