% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{cori_primary}
\alias{cori_primary}
\alias{cori_orange}
\alias{cori_blue}
\alias{cori_purple}
\title{Urban Institute palettes}
\format{
An object of class \code{character} of length 7.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.

An object of class \code{character} of length 2.
}
\source{
Urban Institute Data Visualization Style Guide
\url{http://urbaninstitute.github.io/graphics-styleguide/}
}
\usage{
cori_primary

cori_orange

cori_blue

cori_purple
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the Urban Institute Data Visualization Style Guide.
\url{http://urbaninstitute.github.io/graphics-styleguide/}
}
\concept{urbn palettes}
\keyword{datasets}
