% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{watermark}
\alias{watermark}
\title{Watermark
Cori color scales}
\usage{
watermark(alpha = 0.2)
}
\arguments{
\item{alpha}{Pass the opacity value for the watermark between 0 - 1 (Default: 0.5)}
}
\description{
This is a description
}
