% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_cori.R
\name{scale_color_cori}
\alias{scale_color_cori}
\title{CORI Color Scales}
\usage{
scale_color_cori(color = cori_primary)
}
\arguments{
\item{color}{Pass a named vector color palette. Defaults to \code{cori_primary} colors.}
}
\description{
Implements color palette named vectors as the color scale.
}
