% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_cori.R
\name{scale_fill_cori}
\alias{scale_fill_cori}
\title{Cori Fill scales}
\usage{
scale_fill_cori(color = cori_primary)
}
\arguments{
\item{color}{Pass a cori color palette. Deafults to \code{cori_primary} colors}
}
\description{
This is a description
}
