% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{threshold}
\alias{threshold}
\title{Thresholding}
\usage{
threshold(gg, intercept, axis = "x", label = "", shift = 0)
}
\arguments{
\item{intercept}{x or y intercept (indicated by axis) through which the vertical or horizontal threshold is made.}

\item{axis}{Axis along which a threshold is to be added. Default is "x" indicating a vertical threshold}

\item{label}{Label to add to the graph. Defaults to an empty string}

\item{shift}{Variable to shift the label by "shift" units above thresholding line along "y", or by "shift" units to the right of thresholding line along "x"}
}
\description{
This is a description
}
